+++ src/main/java/com/vaPaTi/vaPaTi/dtos/CampaignStatisticsDTO.java
@@ -6,6 +6,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -16,14 +17,14 @@ public class CampaignStatisticsDTO {
 
     private Long campaignId;
     private String campaignName;
-    private Double amountGoal;
-    private Double amountRaised;
-    private Double percentageReached;
+    private BigDecimal amountGoal;
+    private BigDecimal amountRaised;
+    private BigDecimal percentageReached;
     private Boolean isGoalReached;
     private CampaignStatus status;
     private LocalDateTime targetDate;
     private Long daysRemaining;
     private Long totalDonors;
-    private Double averageDonation;
+    private BigDecimal averageDonation;
     private Long totalDonations;
 }
